Added a "Test for log directory" function.
Added a simple command that will test for the log directory. If the log directory is not there, the "|| " will catch the error and then run a one line function to print a warning to the connections log and then create the directory. This is a modified file for the SSLsplit module.